+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support environment variables to overwrite settings in the `TestAdapter`. (Fix: #57)
 ### Command
 - Update to use `JsonService` implementation to use `Newtonsoft.Json` in the domain.
+### Shared
+- Update to use `ricaun.NamedPipeWrapper.Json`.
 ### TestAdapter
 - Use `LocalExtensionData` to store `TestModel` result.
 - Add `EnvironmentSettings` to overwrite settings in the `TestAdapter`.